browser-tests/pyproject.toml
[project]
name = "verso-blueprint-browser-tests"
version = "0.1.0"
requires-python = ">=3.10"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-playwright",
]
